moved network v1 to separate file
