+++ stackit/internal/services/iaas/network/resource.go
@@ -2,16 +2,9 @@ package network
 
 import (
 	"context"
-	"fmt"
-	"net"
-	"net/http"
-	"strings"
-
-	iaasUtils "github.com/stackitcloud/terraform-provider-stackit/stackit/internal/services/iaas/utils"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/resourcevalidator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
@@ -22,14 +15,21 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
-	"github.com/stackitcloud/stackit-sdk-go/core/oapierror"
 	"github.com/stackitcloud/stackit-sdk-go/services/iaas"
-	"github.com/stackitcloud/stackit-sdk-go/services/iaas/wait"
+	"github.com/stackitcloud/stackit-sdk-go/services/iaasalpha"
 	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/conversion"
 	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/core"
+	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/features"
+	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/services/iaas/network/utils/model"
+	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/services/iaas/network/utils/v1network"
+	iaasUtils "github.com/stackitcloud/terraform-provider-stackit/stackit/internal/services/iaas/utils"
+	iaasAlphaUtils "github.com/stackitcloud/terraform-provider-stackit/stackit/internal/services/iaasalpha/utils"
 	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/utils"
 	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/validate"
 )
+
+// This resource is part of the "network" experiment
+var experiment = "network"
 
 // Ensure the implementation satisfies the expected interfaces.
 var (
@@ -37,30 +37,6 @@ var (
 	_ resource.ResourceWithConfigure   = &networkResource{}
 	_ resource.ResourceWithImportState = &networkResource{}
 )
-
-type Model struct {
-	Id               types.String `tfsdk:"id"` // needed by TF
-	ProjectId        types.String `tfsdk:"project_id"`
-	NetworkId        types.String `tfsdk:"network_id"`
-	Name             types.String `tfsdk:"name"`
-	Nameservers      types.List   `tfsdk:"nameservers"`
-	IPv4Gateway      types.String `tfsdk:"ipv4_gateway"`
-	IPv4Nameservers  types.List   `tfsdk:"ipv4_nameservers"`
-	IPv4Prefix       types.String `tfsdk:"ipv4_prefix"`
-	IPv4PrefixLength types.Int64  `tfsdk:"ipv4_prefix_length"`
-	Prefixes         types.List   `tfsdk:"prefixes"`
-	IPv4Prefixes     types.List   `tfsdk:"ipv4_prefixes"`
-	IPv6Gateway      types.String `tfsdk:"ipv6_gateway"`
-	IPv6Nameservers  types.List   `tfsdk:"ipv6_nameservers"`
-	IPv6Prefix       types.String `tfsdk:"ipv6_prefix"`
-	IPv6PrefixLength types.Int64  `tfsdk:"ipv6_prefix_length"`
-	IPv6Prefixes     types.List   `tfsdk:"ipv6_prefixes"`
-	PublicIP         types.String `tfsdk:"public_ip"`
-	Labels           types.Map    `tfsdk:"labels"`
-	Routed           types.Bool   `tfsdk:"routed"`
-	NoIPv4Gateway    types.Bool   `tfsdk:"no_ipv4_gateway"`
-	NoIPv6Gateway    types.Bool   `tfsdk:"no_ipv6_gateway"`
-}
 
 // NewNetworkResource is a helper function to simplify the provider implementation.
 func NewNetworkResource() resource.Resource {
@@ -70,6 +46,10 @@ func NewNetworkResource() resource.Resource {
 // networkResource is the resource implementation.
 type networkResource struct {
 	client *iaas.APIClient
+	// alphaClient will be used in case the experimental flag "network" is set
+	alphaClient    *iaasalpha.APIClient
+	isExperimental bool
+	providerData   core.ProviderData
 }
 
 // Metadata returns the resource type name.
@@ -79,27 +59,72 @@ func (r *networkResource) Metadata(_ context.Context, req resource.MetadataReque
 
 // Configure adds the provider configured client to the resource.
 func (r *networkResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
-	providerData, ok := conversion.ParseProviderData(ctx, req.ProviderData, &resp.Diagnostics)
+	var ok bool
+	r.providerData, ok = conversion.ParseProviderData(ctx, req.ProviderData, &resp.Diagnostics)
 	if !ok {
 		return
 	}
 
-	apiClient := iaasUtils.ConfigureClient(ctx, &providerData, &resp.Diagnostics)
+	// TODO: check if this works like intended
+	r.isExperimental = features.CheckExperimentEnabledWithoutError(ctx, &r.providerData, experiment, "stackit_network", &resp.Diagnostics)
 	if resp.Diagnostics.HasError() {
 		return
 	}
-	r.client = apiClient
+
+	if r.isExperimental {
+		alphaApiClient := iaasAlphaUtils.ConfigureClient(ctx, &r.providerData, &resp.Diagnostics)
+		if resp.Diagnostics.HasError() {
+			return
+		}
+		r.alphaClient = alphaApiClient
+	} else {
+		apiClient := iaasUtils.ConfigureClient(ctx, &r.providerData, &resp.Diagnostics)
+		if resp.Diagnostics.HasError() {
+			return
+		}
+		r.client = apiClient
+	}
 	tflog.Info(ctx, "IaaS client configured")
 }
 
-func (r networkResource) ValidateConfig(ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse) {
-	var model Model
-	resp.Diagnostics.Append(req.Config.Get(ctx, &model)...)
+// ModifyPlan implements resource.ResourceWithModifyPlan.
+// Use the modifier to set the effective region in the current plan.
+func (r *networkResource) ModifyPlan(ctx context.Context, req resource.ModifyPlanRequest, resp *resource.ModifyPlanResponse) { // nolint:gocritic // function signature required by Terraform
+	var configModel model.Model
+	// skip initial empty configuration to avoid follow-up errors
+	if req.Config.Raw.IsNull() {
+		return
+	}
+	resp.Diagnostics.Append(req.Config.Get(ctx, &configModel)...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	if !model.Nameservers.IsUnknown() && !model.IPv4Nameservers.IsUnknown() && !model.Nameservers.IsNull() && !model.IPv4Nameservers.IsNull() {
+	var planModel model.Model
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &planModel)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	utils.AdaptRegion(ctx, configModel.Region, &planModel.Region, r.providerData.GetRegion(), resp)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	resp.Diagnostics.Append(resp.Plan.Set(ctx, planModel)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+}
+
+func (r *networkResource) ValidateConfig(ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse) {
+	var resourceModel model.Model
+	resp.Diagnostics.Append(req.Config.Get(ctx, &resourceModel)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if !resourceModel.Nameservers.IsUnknown() && !resourceModel.IPv4Nameservers.IsUnknown() && !resourceModel.Nameservers.IsNull() && !resourceModel.IPv4Nameservers.IsNull() {
 		core.LogAndAddError(ctx, &resp.Diagnostics, "Error configuring network", "You cannot provide both the `nameservers` and `ipv4_nameservers` fields simultaneously. Please remove the deprecated `nameservers` field, and use `ipv4_nameservers` to configure nameservers for IPv4.")
 	}
 }
@@ -285,498 +310,57 @@ func (r *networkResource) Schema(_ context.Context, _ resource.SchemaRequest, re
 					boolplanmodifier.RequiresReplace(),
 				},
 			},
+			"region": schema.StringAttribute{
+				Optional: true,
+				// must be computed to allow for storing the override value from the provider
+				Computed:    true,
+				Description: "Can only be used when experimental \"network\" is set.\nThe resource region. If not defined, the provider region is used.",
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
+				},
+			},
 		},
 	}
 }
 
 // Create creates the resource and sets the initial Terraform state.
 func (r *networkResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) { // nolint:gocritic // function signature required by Terraform
-	// Retrieve values from plan
-	var model Model
-	diags := req.Plan.Get(ctx, &model)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
+	if !r.isExperimental {
+		v1network.Create(ctx, req, resp, r.client)
 		return
 	}
-
-	projectId := model.ProjectId.ValueString()
-	ctx = tflog.SetField(ctx, "project_id", projectId)
-
-	// Generate API request body from model
-	payload, err := toCreatePayload(ctx, &model)
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error creating network", fmt.Sprintf("Creating API payload: %v", err))
-		return
-	}
-
-	// Create new network
-
-	network, err := r.client.CreateNetwork(ctx, projectId).CreateNetworkPayload(*payload).Execute()
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error creating network", fmt.Sprintf("Calling API: %v", err))
-		return
-	}
-
-	networkId := *network.NetworkId
-	network, err = wait.CreateNetworkWaitHandler(ctx, r.client, projectId, networkId).WaitWithContext(ctx)
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error creating network", fmt.Sprintf("Network creation waiting: %v", err))
-		return
-	}
-
-	ctx = tflog.SetField(ctx, "network_id", networkId)
-
-	// Map response body to schema
-	err = mapFields(ctx, network, &model)
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error creating network", fmt.Sprintf("Processing API payload: %v", err))
-		return
-	}
-	// Set state to fully populated data
-	diags = resp.State.Set(ctx, model)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-	tflog.Info(ctx, "Network created")
+	// TODO: Add v2alpha
 }
 
 // Read refreshes the Terraform state with the latest data.
 func (r *networkResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) { // nolint:gocritic // function signature required by Terraform
-	var model Model
-	diags := req.State.Get(ctx, &model)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
+	if !r.isExperimental {
+		v1network.Read(ctx, req, resp, r.client)
 	}
-	projectId := model.ProjectId.ValueString()
-	networkId := model.NetworkId.ValueString()
-	ctx = tflog.SetField(ctx, "project_id", projectId)
-	ctx = tflog.SetField(ctx, "network_id", networkId)
-
-	networkResp, err := r.client.GetNetwork(ctx, projectId, networkId).Execute()
-	if err != nil {
-		oapiErr, ok := err.(*oapierror.GenericOpenAPIError) //nolint:errorlint //complaining that error.As should be used to catch wrapped errors, but this error should not be wrapped
-		if ok && oapiErr.StatusCode == http.StatusNotFound {
-			resp.State.RemoveResource(ctx)
-			return
-		}
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error reading network", fmt.Sprintf("Calling API: %v", err))
-		return
-	}
-
-	// Map response body to schema
-	err = mapFields(ctx, networkResp, &model)
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error reading network", fmt.Sprintf("Processing API payload: %v", err))
-		return
-	}
-	// Set refreshed state
-	diags = resp.State.Set(ctx, model)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-	tflog.Info(ctx, "Network read")
+	// TODO: Add v2alpha
 }
 
 // Update updates the resource and sets the updated Terraform state on success.
 func (r *networkResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) { // nolint:gocritic // function signature required by Terraform
-	// Retrieve values from plan
-	var model Model
-	diags := req.Plan.Get(ctx, &model)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
+	if !r.isExperimental {
+		v1network.Update(ctx, req, resp, r.client)
 	}
-	projectId := model.ProjectId.ValueString()
-	networkId := model.NetworkId.ValueString()
-	ctx = tflog.SetField(ctx, "project_id", projectId)
-	ctx = tflog.SetField(ctx, "network_id", networkId)
-
-	// Retrieve values from state
-	var stateModel Model
-	diags = req.State.Get(ctx, &stateModel)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	// Generate API request body from model
-	payload, err := toUpdatePayload(ctx, &model, &stateModel)
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error updating network", fmt.Sprintf("Creating API payload: %v", err))
-		return
-	}
-	// Update existing network
-	err = r.client.PartialUpdateNetwork(ctx, projectId, networkId).PartialUpdateNetworkPayload(*payload).Execute()
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error updating network", fmt.Sprintf("Calling API: %v", err))
-		return
-	}
-	waitResp, err := wait.UpdateNetworkWaitHandler(ctx, r.client, projectId, networkId).WaitWithContext(ctx)
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error updating network", fmt.Sprintf("Network update waiting: %v", err))
-		return
-	}
-
-	err = mapFields(ctx, waitResp, &model)
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error updating network", fmt.Sprintf("Processing API payload: %v", err))
-		return
-	}
-	diags = resp.State.Set(ctx, model)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-	tflog.Info(ctx, "Network updated")
+	// TODO: Add v2alpha
 }
 
 // Delete deletes the resource and removes the Terraform state on success.
 func (r *networkResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) { // nolint:gocritic // function signature required by Terraform
-	// Retrieve values from state
-	var model Model
-	diags := req.State.Get(ctx, &model)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
+	if !r.isExperimental {
+		v1network.Delete(ctx, req, resp, r.client)
 	}
-
-	projectId := model.ProjectId.ValueString()
-	networkId := model.NetworkId.ValueString()
-	ctx = tflog.SetField(ctx, "project_id", projectId)
-	ctx = tflog.SetField(ctx, "network_id", networkId)
-
-	// Delete existing network
-	err := r.client.DeleteNetwork(ctx, projectId, networkId).Execute()
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error deleting network", fmt.Sprintf("Calling API: %v", err))
-		return
-	}
-	_, err = wait.DeleteNetworkWaitHandler(ctx, r.client, projectId, networkId).WaitWithContext(ctx)
-	if err != nil {
-		core.LogAndAddError(ctx, &resp.Diagnostics, "Error deleting network", fmt.Sprintf("Network deletion waiting: %v", err))
-		return
-	}
-
-	tflog.Info(ctx, "Network deleted")
+	// TODO: Add v2alpha
 }
 
 // ImportState imports a resource into the Terraform state on success.
 // The expected format of the resource import identifier is: project_id,network_id
 func (r *networkResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
-	idParts := strings.Split(req.ID, core.Separator)
-
-	if len(idParts) != 2 || idParts[0] == "" || idParts[1] == "" {
-		core.LogAndAddError(ctx, &resp.Diagnostics,
-			"Error importing network",
-			fmt.Sprintf("Expected import identifier with format: [project_id],[network_id]  Got: %q", req.ID),
-		)
-		return
+	if !r.isExperimental {
+		v1network.ImportState(ctx, req, resp)
 	}
-
-	projectId := idParts[0]
-	networkId := idParts[1]
-	ctx = tflog.SetField(ctx, "project_id", projectId)
-	ctx = tflog.SetField(ctx, "network_id", networkId)
-
-	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("project_id"), projectId)...)
-	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("network_id"), networkId)...)
-	tflog.Info(ctx, "Network state imported")
-}
-
-func mapFields(ctx context.Context, networkResp *iaas.Network, model *Model) error {
-	if networkResp == nil {
-		return fmt.Errorf("response input is nil")
-	}
-	if model == nil {
-		return fmt.Errorf("model input is nil")
-	}
-
-	var networkId string
-	if model.NetworkId.ValueString() != "" {
-		networkId = model.NetworkId.ValueString()
-	} else if networkResp.NetworkId != nil {
-		networkId = *networkResp.NetworkId
-	} else {
-		return fmt.Errorf("network id not present")
-	}
-
-	model.Id = utils.BuildInternalTerraformId(model.ProjectId.ValueString(), networkId)
-
-	labels, err := iaasUtils.MapLabels(ctx, networkResp.Labels, model.Labels)
-	if err != nil {
-		return err
-	}
-
-	// IPv4
-	if networkResp.Nameservers == nil {
-		model.Nameservers = types.ListNull(types.StringType)
-		model.IPv4Nameservers = types.ListNull(types.StringType)
-	} else {
-		respNameservers := *networkResp.Nameservers
-		modelNameservers, err := utils.ListValuetoStringSlice(model.Nameservers)
-		modelIPv4Nameservers, errIpv4 := utils.ListValuetoStringSlice(model.IPv4Nameservers)
-		if err != nil {
-			return fmt.Errorf("get current network nameservers from model: %w", err)
-		}
-		if errIpv4 != nil {
-			return fmt.Errorf("get current IPv4 network nameservers from model: %w", errIpv4)
-		}
-
-		reconciledNameservers := utils.ReconcileStringSlices(modelNameservers, respNameservers)
-		reconciledIPv4Nameservers := utils.ReconcileStringSlices(modelIPv4Nameservers, respNameservers)
-
-		nameserversTF, diags := types.ListValueFrom(ctx, types.StringType, reconciledNameservers)
-		ipv4NameserversTF, ipv4Diags := types.ListValueFrom(ctx, types.StringType, reconciledIPv4Nameservers)
-		if diags.HasError() {
-			return fmt.Errorf("map network nameservers: %w", core.DiagsToError(diags))
-		}
-		if ipv4Diags.HasError() {
-			return fmt.Errorf("map IPv4 network nameservers: %w", core.DiagsToError(ipv4Diags))
-		}
-
-		model.Nameservers = nameserversTF
-		model.IPv4Nameservers = ipv4NameserversTF
-	}
-
-	if networkResp.Prefixes == nil {
-		model.Prefixes = types.ListNull(types.StringType)
-		model.IPv4Prefixes = types.ListNull(types.StringType)
-	} else {
-		respPrefixes := *networkResp.Prefixes
-		prefixesTF, diags := types.ListValueFrom(ctx, types.StringType, respPrefixes)
-		if diags.HasError() {
-			return fmt.Errorf("map network prefixes: %w", core.DiagsToError(diags))
-		}
-		if len(respPrefixes) > 0 {
-			model.IPv4Prefix = types.StringValue(respPrefixes[0])
-			_, netmask, err := net.ParseCIDR(respPrefixes[0])
-			if err != nil {
-				// silently ignore parsing error for the netmask
-				model.IPv4PrefixLength = types.Int64Null()
-			} else {
-				ones, _ := netmask.Mask.Size()
-				model.IPv4PrefixLength = types.Int64Value(int64(ones))
-			}
-		}
-
-		model.Prefixes = prefixesTF
-		model.IPv4Prefixes = prefixesTF
-	}
-
-	if networkResp.Gateway != nil {
-		model.IPv4Gateway = types.StringPointerValue(networkResp.GetGateway())
-	} else {
-		model.IPv4Gateway = types.StringNull()
-	}
-
-	// IPv6
-
-	if networkResp.NameserversV6 == nil {
-		model.IPv6Nameservers = types.ListNull(types.StringType)
-	} else {
-		respIPv6Nameservers := *networkResp.NameserversV6
-		modelIPv6Nameservers, errIpv6 := utils.ListValuetoStringSlice(model.IPv6Nameservers)
-		if errIpv6 != nil {
-			return fmt.Errorf("get current IPv6 network nameservers from model: %w", errIpv6)
-		}
-
-		reconciledIPv6Nameservers := utils.ReconcileStringSlices(modelIPv6Nameservers, respIPv6Nameservers)
-
-		ipv6NameserversTF, ipv6Diags := types.ListValueFrom(ctx, types.StringType, reconciledIPv6Nameservers)
-		if ipv6Diags.HasError() {
-			return fmt.Errorf("map IPv6 network nameservers: %w", core.DiagsToError(ipv6Diags))
-		}
-
-		model.IPv6Nameservers = ipv6NameserversTF
-	}
-
-	if networkResp.PrefixesV6 == nil {
-		model.IPv6Prefixes = types.ListNull(types.StringType)
-	} else {
-		respPrefixesV6 := *networkResp.PrefixesV6
-		prefixesV6TF, diags := types.ListValueFrom(ctx, types.StringType, respPrefixesV6)
-		if diags.HasError() {
-			return fmt.Errorf("map network IPv6 prefixes: %w", core.DiagsToError(diags))
-		}
-		if len(respPrefixesV6) > 0 {
-			model.IPv6Prefix = types.StringValue(respPrefixesV6[0])
-			_, netmask, err := net.ParseCIDR(respPrefixesV6[0])
-			if err != nil {
-				// silently ignore parsing error for the netmask
-				model.IPv6PrefixLength = types.Int64Null()
-			} else {
-				ones, _ := netmask.Mask.Size()
-				model.IPv6PrefixLength = types.Int64Value(int64(ones))
-			}
-		}
-		model.IPv6Prefixes = prefixesV6TF
-	}
-
-	if networkResp.Gatewayv6 != nil {
-		model.IPv6Gateway = types.StringPointerValue(networkResp.GetGatewayv6())
-	} else {
-		model.IPv6Gateway = types.StringNull()
-	}
-
-	model.NetworkId = types.StringValue(networkId)
-	model.Name = types.StringPointerValue(networkResp.Name)
-	model.PublicIP = types.StringPointerValue(networkResp.PublicIp)
-	model.Labels = labels
-	model.Routed = types.BoolPointerValue(networkResp.Routed)
-
-	return nil
-}
-
-func toCreatePayload(ctx context.Context, model *Model) (*iaas.CreateNetworkPayload, error) {
-	if model == nil {
-		return nil, fmt.Errorf("nil model")
-	}
-	addressFamily := &iaas.CreateNetworkAddressFamily{}
-
-	modelIPv6Nameservers := []string{}
-	for _, ipv6ns := range model.IPv6Nameservers.Elements() {
-		ipv6NameserverString, ok := ipv6ns.(types.String)
-		if !ok {
-			return nil, fmt.Errorf("type assertion failed")
-		}
-		modelIPv6Nameservers = append(modelIPv6Nameservers, ipv6NameserverString.ValueString())
-	}
-
-	if !(model.IPv6Prefix.IsNull() || model.IPv6PrefixLength.IsNull() || model.IPv6Nameservers.IsNull()) {
-		addressFamily.Ipv6 = &iaas.CreateNetworkIPv6Body{
-			Nameservers:  &modelIPv6Nameservers,
-			Prefix:       conversion.StringValueToPointer(model.IPv6Prefix),
-			PrefixLength: conversion.Int64ValueToPointer(model.IPv6PrefixLength),
-		}
-
-		if model.NoIPv6Gateway.ValueBool() {
-			addressFamily.Ipv6.Gateway = iaas.NewNullableString(nil)
-		} else if !(model.IPv6Gateway.IsUnknown() || model.IPv6Gateway.IsNull()) {
-			addressFamily.Ipv6.Gateway = iaas.NewNullableString(conversion.StringValueToPointer(model.IPv6Gateway))
-		}
-	}
-
-	modelIPv4Nameservers := []string{}
-	var modelIPv4List []attr.Value
-
-	if !(model.IPv4Nameservers.IsNull() || model.IPv4Nameservers.IsUnknown()) {
-		modelIPv4List = model.IPv4Nameservers.Elements()
-	} else {
-		modelIPv4List = model.Nameservers.Elements()
-	}
-
-	for _, ipv4ns := range modelIPv4List {
-		ipv4NameserverString, ok := ipv4ns.(types.String)
-		if !ok {
-			return nil, fmt.Errorf("type assertion failed")
-		}
-		modelIPv4Nameservers = append(modelIPv4Nameservers, ipv4NameserverString.ValueString())
-	}
-
-	if !model.IPv4Prefix.IsNull() || !model.IPv4PrefixLength.IsNull() || !model.IPv4Nameservers.IsNull() || !model.Nameservers.IsNull() {
-		addressFamily.Ipv4 = &iaas.CreateNetworkIPv4Body{
-			Nameservers:  &modelIPv4Nameservers,
-			Prefix:       conversion.StringValueToPointer(model.IPv4Prefix),
-			PrefixLength: conversion.Int64ValueToPointer(model.IPv4PrefixLength),
-		}
-
-		if model.NoIPv4Gateway.ValueBool() {
-			addressFamily.Ipv4.Gateway = iaas.NewNullableString(nil)
-		} else if !(model.IPv4Gateway.IsUnknown() || model.IPv4Gateway.IsNull()) {
-			addressFamily.Ipv4.Gateway = iaas.NewNullableString(conversion.StringValueToPointer(model.IPv4Gateway))
-		}
-	}
-
-	labels, err := conversion.ToStringInterfaceMap(ctx, model.Labels)
-	if err != nil {
-		return nil, fmt.Errorf("converting to Go map: %w", err)
-	}
-
-	payload := iaas.CreateNetworkPayload{
-		Name:   conversion.StringValueToPointer(model.Name),
-		Labels: &labels,
-		Routed: conversion.BoolValueToPointer(model.Routed),
-	}
-
-	if addressFamily.Ipv6 != nil || addressFamily.Ipv4 != nil {
-		payload.AddressFamily = addressFamily
-	}
-
-	return &payload, nil
-}
-
-func toUpdatePayload(ctx context.Context, model, stateModel *Model) (*iaas.PartialUpdateNetworkPayload, error) {
-	if model == nil {
-		return nil, fmt.Errorf("nil model")
-	}
-	addressFamily := &iaas.UpdateNetworkAddressFamily{}
-
-	modelIPv6Nameservers := []string{}
-	for _, ipv6ns := range model.IPv6Nameservers.Elements() {
-		ipv6NameserverString, ok := ipv6ns.(types.String)
-		if !ok {
-			return nil, fmt.Errorf("type assertion failed")
-		}
-		modelIPv6Nameservers = append(modelIPv6Nameservers, ipv6NameserverString.ValueString())
-	}
-
-	if !(model.IPv6Nameservers.IsNull() || model.IPv6Nameservers.IsUnknown()) {
-		addressFamily.Ipv6 = &iaas.UpdateNetworkIPv6Body{
-			Nameservers: &modelIPv6Nameservers,
-		}
-
-		if model.NoIPv6Gateway.ValueBool() {
-			addressFamily.Ipv6.Gateway = iaas.NewNullableString(nil)
-		} else if !(model.IPv6Gateway.IsUnknown() || model.IPv6Gateway.IsNull()) {
-			addressFamily.Ipv6.Gateway = iaas.NewNullableString(conversion.StringValueToPointer(model.IPv6Gateway))
-		}
-	}
-
-	modelIPv4Nameservers := []string{}
-	var modelIPv4List []attr.Value
-
-	if !(model.IPv4Nameservers.IsNull() || model.IPv4Nameservers.IsUnknown()) {
-		modelIPv4List = model.IPv4Nameservers.Elements()
-	} else {
-		modelIPv4List = model.Nameservers.Elements()
-	}
-	for _, ipv4ns := range modelIPv4List {
-		ipv4NameserverString, ok := ipv4ns.(types.String)
-		if !ok {
-			return nil, fmt.Errorf("type assertion failed")
-		}
-		modelIPv4Nameservers = append(modelIPv4Nameservers, ipv4NameserverString.ValueString())
-	}
-
-	if !model.IPv4Nameservers.IsNull() || !model.Nameservers.IsNull() {
-		addressFamily.Ipv4 = &iaas.UpdateNetworkIPv4Body{
-			Nameservers: &modelIPv4Nameservers,
-		}
-
-		if model.NoIPv4Gateway.ValueBool() {
-			addressFamily.Ipv4.Gateway = iaas.NewNullableString(nil)
-		} else if !(model.IPv4Gateway.IsUnknown() || model.IPv4Gateway.IsNull()) {
-			addressFamily.Ipv4.Gateway = iaas.NewNullableString(conversion.StringValueToPointer(model.IPv4Gateway))
-		}
-	}
-
-	currentLabels := stateModel.Labels
-	labels, err := conversion.ToJSONMapPartialUpdatePayload(ctx, currentLabels, model.Labels)
-	if err != nil {
-		return nil, fmt.Errorf("converting to Go map: %w", err)
-	}
-
-	payload := iaas.PartialUpdateNetworkPayload{
-		Name:   conversion.StringValueToPointer(model.Name),
-		Labels: &labels,
-	}
-
-	if addressFamily.Ipv6 != nil || addressFamily.Ipv4 != nil {
-		payload.AddressFamily = addressFamily
-	}
-
-	return &payload, nil
+	// TODO: Add v2alpha
 }
