+++ stackit/internal/services/iaas/network/utils/model/model.go
@@ -1,0 +1,28 @@
+package model
+
+import "github.com/hashicorp/terraform-plugin-framework/types"
+
+type Model struct {
+	Id               types.String `tfsdk:"id"` // needed by TF
+	ProjectId        types.String `tfsdk:"project_id"`
+	NetworkId        types.String `tfsdk:"network_id"`
+	Name             types.String `tfsdk:"name"`
+	Nameservers      types.List   `tfsdk:"nameservers"`
+	IPv4Gateway      types.String `tfsdk:"ipv4_gateway"`
+	IPv4Nameservers  types.List   `tfsdk:"ipv4_nameservers"`
+	IPv4Prefix       types.String `tfsdk:"ipv4_prefix"`
+	IPv4PrefixLength types.Int64  `tfsdk:"ipv4_prefix_length"`
+	Prefixes         types.List   `tfsdk:"prefixes"`
+	IPv4Prefixes     types.List   `tfsdk:"ipv4_prefixes"`
+	IPv6Gateway      types.String `tfsdk:"ipv6_gateway"`
+	IPv6Nameservers  types.List   `tfsdk:"ipv6_nameservers"`
+	IPv6Prefix       types.String `tfsdk:"ipv6_prefix"`
+	IPv6PrefixLength types.Int64  `tfsdk:"ipv6_prefix_length"`
+	IPv6Prefixes     types.List   `tfsdk:"ipv6_prefixes"`
+	PublicIP         types.String `tfsdk:"public_ip"`
+	Labels           types.Map    `tfsdk:"labels"`
+	Routed           types.Bool   `tfsdk:"routed"`
+	NoIPv4Gateway    types.Bool   `tfsdk:"no_ipv4_gateway"`
+	NoIPv6Gateway    types.Bool   `tfsdk:"no_ipv6_gateway"`
+	Region           types.String `tfsdk:"region"`
+}
